+++ test/_support/test_template.ex
@@ -10,19 +10,20 @@ defmodule SpectorTest.TestTemplate do
         end
       end)
 
-    full_code = quote do
-      defmodule unquote(module) do
-        use ExUnit.Case
+    full_code =
+      quote do
+        defmodule unquote(module) do
+          use ExUnit.Case
 
-        setup do
-          :ok = Ecto.Adapters.SQL.Sandbox.checkout(SpectorTest.Repo)
-        end
+          setup do
+            :ok = Ecto.Adapters.SQL.Sandbox.checkout(SpectorTest.Repo)
+          end
 
-        test "usage examples" do
-          unquote(combined)
+          test "usage examples" do
+            unquote(combined)
+          end
         end
       end
-    end
 
     Code.eval_quoted(full_code)
   end